Remove hosted background service code
