+++ ExcelReader.tonyissa/ExcelReader/appsettings.json
@@ -1,5 +1,5 @@
 {
   "ConnectionStrings": {
-    "DefaultConnection": "Server=(localdb)\\TheCsharpAcademyProjects;Database=ExcelReader;"
+    "DefaultConnection": "Data Source=excel-reader.db"
   }
 }
